Separate 'loader' utility class from ContextManagers.
